changed Manager to be stack object per backupStart #10
Now the manager state singleton is in a static member _currentManager
that is set when we get our first poll callback (only the error callback
is called when you fail because another backup is running).  This is
protected by _currentMutex.

There may be a small race possibility during the period after the backup
has finished and released its internal lock and before the Manager is
destroyed (and resets _currentManager to NULL) when, if another backup
starts and gets its first poll we can conflict.  This is noted in a
comment.

closes #10